Add files via upload
This takes two input files - one with GP surgery data and another with prescription data, and produces a plot of prescription rate vs the range (n_distinct) of medications prescribed.  The data are then fitted to a reflected exponential and compared to two simulations: one with a uniform distribution and another using the measured frequency distribution of the medications from the same master dataset.